Refactoring and comments for interface functions
Minor refactoring of the scene object handling code.

Changed several functions, mostly in the geom library, to return const
values when appropriate.

Added comments describing some of the public interface functions for some
of the scene and manager classes.

+++ src/managers/ServiceMgr.cpp
@@ -70,7 +70,7 @@ void ServiceMgr::registerSystem(Systems::AbstractSystem *sys) {
 
 
 Systems::AbstractSystem* ServiceMgr::getSystem(int sysid) {
-    if (sysid < 0 || sysid >= p_systems.size())
+    if (sysid < 0 || sysid >= (int) p_systems.size())
         return nullptr;
 
     return p_systems.at(sysid);
